requirements.txt
albumentations==1.3.0
asttokens==2.1.0
av==10.0.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
customtkinter==5.0.3
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.3
decorator==5.1.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
ffmpeg-python==0.2.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
imageio==2.23.0
ipykernel==6.17.0
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter_client==7.4.4
jupyter_core==4.11.2
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.4
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.68
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.3.0
pretrainedmodels==0.7.4
promise==2.3
prompt-toolkit==3.0.32
protobuf==4.21.12
psutil==5.9.3
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-vlc==3.0.18121
pytz==2022.6
PyWavelets==1.4.1
pywin32==304
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
segmentation-models-pytorch==0.3.0
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
stack-data==0.6.0
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.4.12
tkvideoplayer==2.3
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.7
wcwidth==0.2.5
XlsxWriter==3.0.5
diskcache==5.4.0